pyproject.toml
[tool.poetry]
name = "insurance-api"
version = "0.1.0"
description = ""
authors = ["sxw111 <sxw9sxw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {version = "^0.32.1", extras = ["standart"]}
sqlalchemy = {version = "^2.0.36", extras = ["asyncio"]}
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
aiokafka = "^0.12.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
